Update README.md
DESIGN AND SIMULATION OF THREE PHASE TRAFFIC LIGHT MONITORING AND ALARMING SYSTEM FOR TRAFFIC OFFENDER 
This project offer a modern way through the use of infrared proximity sensors, will detect vehicles passing through a specific lane when traffic lights do not dictate movement to this lane. So, the traffic conditions on the road can be known and predicted.